@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import { Route, Switch, Redirect } from 'react-router-dom'
+import { Route, Routes, Navigate } from 'react-router-dom'
 import { Teleop } from '@/renderer/components/pages/Teleop'
 import { Victim } from '@/renderer/components/pages/Victim'
 import { Debug } from '@/renderer/components/pages/Debug'
@@ -8,14 +8,13 @@ import { ConfigPage } from '@/renderer/components/pages/Config/ConfigPage'
 export const Router: FC = () => {
   return (
     <>
-      <Redirect exact from="/" to="/teleop" />
-
-      <Switch>
-        <Route path="/teleop" component={Teleop} />
-        <Route path="/victim" component={Victim} />
-        <Route path="/config" component={ConfigPage} />
-        <Route path="/debug" component={Debug} />
-      </Switch>
+      <Routes>
+        <Route path="/" element={<Navigate to="/teleop" />} />
+        <Route path="/teleop" element={<Teleop />} />
+        <Route path="/victim" element={<Victim />} />
+        <Route path="/config/*" element={<ConfigPage />} />
+        <Route path="/debug" element={<Debug />} />
+      </Routes>
     </>
   )
 }
