@@ -6,7 +6,7 @@ import {
   AUDIO_STOP,
 } from '@/shared/constants'
 import { ipcMain, IpcMainEvent } from 'electron'
-import execa, { ExecaChildProcess } from 'execa'
+import { execa, ExecaChildProcess } from 'execa'
 
 // This is a pretty hacky solution to have 2 way audio
 // Essentially we don't use electron and let ros handle it.
